requirements.txt
scikit-image
opencv-python
PyYAML
albumentations
pytorch-metric-learning
mxnet